Implement fuzz_station()
This function has been used for testing various function numbers
(various values of byte at position 0x07).
It was checked whether the weather station would respond to these
packets and how.

This way, the function 0x90 was discovered ("reply instantly with
sensor data"), among other things.

This function was used with some dirty hacks in main(), which were not
suitable for being kept in the program.